Work on tooltips
Need to figure out how to create a div or something when the user hovers over a table row, that then displays that table rows relevant blurb and item name. Sort of works. Console logs show the right data is being passed through, but not rendering anything. Try and solve that. Maybe make a blank div show up with a bright yellow border above the position of the element its called from to at least see it working.
Also updated file structure. Components now live in major or minor folders, depending on how heavy they are